+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - TF-IDF Transformer is now Reversible
 
 1.0.1
+    - Prevent certain specification false negatives
 
 - 1.0.0
     - No changes